Fix for bug #12.
Bug#12 is to address the library not working for the Ardino
Zero boards. The issue seems to be that for this board, using IDE 1.6.9,
the ports now return a uint32_t instead f a uint8_t.

To address this issue, I added a marco case to define the types of the
PortReg and Port Mask to be uint32_t. I chose a macro that was clearly
defined in the variant.h of the zero.